requirements.txt
PyQt6>=6.4.0
psutil>=5.9.0
pywin32>=305
wmi>=1.5.1
speedtest-cli>=2.1.3
qt-material>=2.14
Pillow>=10.0.0
